Fix Copilot review: remove duplicated helpers, unused imports, align mock schema
- Extract registerMockBookingsFeed to shared helper in xstate-mocks.ts
- Remove ~340-line duplicated function from both full-flow test files
- Fix noShowedAt: use null instead of zeroTimestamp
- Fix machineId: use "MC Booking Request" matching actual xstate config
- Add missing "use client" directive to edit/role/[id]/page.tsx
