+++ booking-app/tests/e2e/xstate-service-closeout-full-flow.e2e.test.ts
@@ -12,11 +12,8 @@ import {
 import { registerBookingMocks } from "./helpers/mock-routes";
 import {
   JSON_HEADERS,
-  TIMESTAMP_FIELDS_WITH_BY,
   createTimestamp,
-  serializeBookingRecord,
-  serializeGenericRecord,
-  registerDefinePropertyInterceptor,
+  registerMockBookingsFeed,
 } from "./helpers/xstate-mocks";
 
 const BASE_URL =
@@ -248,347 +245,6 @@ async function seedCheckedOutBookingWithAllServices() {
   });
 }
 
-async function registerMockBookingsFeed(page: Page) {
-  const snapshot = await getDoc(
-    doc({} as any, "mc-bookings", BOOKING_DOC_ID)
-  );
-  const baseData = snapshot.data();
-  const basePayload = baseData
-    ? {
-        id: BOOKING_DOC_ID,
-        calendarEventId: CALENDAR_EVENT_ID,
-        ...serializeBookingRecord(baseData),
-      }
-    : null;
-
-  const rightsSnapshot = await getDoc(
-    doc({} as any, "mc-usersRights", USERS_RIGHTS_DOC_ID)
-  );
-  const rightsPayload = rightsSnapshot.data()
-    ? {
-        id: USERS_RIGHTS_DOC_ID,
-        ...serializeGenericRecord(rightsSnapshot.data()),
-      }
-    : null;
-
-  await page.route("**/api/__mock__/bookings", async (route) => {
-    const latestSnapshot = await getDoc(
-      doc({} as any, "mc-bookings", BOOKING_DOC_ID)
-    );
-    const latestData = latestSnapshot.data();
-    const payload = latestData
-      ? [
-          {
-            id: BOOKING_DOC_ID,
-            calendarEventId: CALENDAR_EVENT_ID,
-            ...serializeBookingRecord(latestData),
-          },
-        ]
-      : [];
-
-    await route.fulfill({
-      status: 200,
-      headers: JSON_HEADERS,
-      body: JSON.stringify(payload),
-    });
-  });
-
-  await registerDefinePropertyInterceptor(page);
-
-  await page.addInitScript(
-    ({
-      timestampFields,
-      initialBooking,
-      initialUsersRights,
-      adminEmail,
-    }) => {
-      const makeTimestamp = (value: string | null | undefined) => {
-        if (!value) return null;
-        const baseDate = new Date(value);
-        return {
-          toDate: () => new Date(baseDate),
-          toMillis: () => baseDate.getTime(),
-          valueOf: () => baseDate.getTime(),
-        };
-      };
-
-      const enrichBooking = (raw: Record<string, any>) => {
-        const booking = { ...raw };
-        timestampFields.forEach((field: string) => {
-          if (booking[field]) {
-            booking[field] = makeTimestamp(booking[field]);
-          } else if (
-            booking[field] === null ||
-            booking[field] === undefined
-          ) {
-            booking[field] = null;
-          }
-        });
-        if (booking.startDate) {
-          booking.startDate = makeTimestamp(booking.startDate);
-        }
-        if (booking.endDate) {
-          booking.endDate = makeTimestamp(booking.endDate);
-        }
-        if (booking.requestedAt) {
-          booking.requestedAt = makeTimestamp(booking.requestedAt);
-        }
-        return booking;
-      };
-
-      async function loadBookings() {
-        const response = await fetch("/api/__mock__/bookings");
-        const raw = await response.json();
-        const bookings = raw.map(enrichBooking);
-        (window as any).__mockBookings = bookings;
-        (window as any).__bookingE2EMocks = {
-          bookings,
-          usersRights: (initialUsersRights ? [initialUsersRights] : []).map(
-            (record: any) => {
-              const enriched = { ...record };
-              ["createdAt", "updatedAt"].forEach((field) => {
-                if (enriched[field]) {
-                  enriched[field] = makeTimestamp(enriched[field]);
-                }
-              });
-              return enriched;
-            }
-          ),
-          usersApprovers: [
-            {
-              email: adminEmail,
-              department: "ITP",
-              level: 3,
-              createdAt: new Date().toISOString(),
-            },
-          ],
-        };
-        return bookings;
-      }
-
-      async function ensureBookings() {
-        if (!(window as any).__mockBookings) {
-          return await loadBookings();
-        }
-        return (window as any).__mockBookings;
-      }
-
-      (window as any).__refreshMockBookings = async () => {
-        return await loadBookings();
-      };
-
-      const originalClientFetch =
-        (window as any).clientFetchAllDataFromCollection;
-
-      const usersRightsRecords = (
-        initialUsersRights ? [initialUsersRights] : []
-      ).map((record: any) => {
-        const enriched = { ...record };
-        ["createdAt", "updatedAt"].forEach((field) => {
-          if (enriched[field]) {
-            enriched[field] = makeTimestamp(enriched[field]);
-          }
-        });
-        return enriched;
-      });
-
-      (window as any).__mockUsersRights = usersRightsRecords;
-
-      (window as any).clientFetchAllDataFromCollection = async function (
-        tableName: string,
-        constraints: unknown[],
-        tenant: string
-      ) {
-        const normalizedTableName = tableName
-          ? tableName.toLowerCase()
-          : "";
-
-        if (
-          normalizedTableName.includes("booking") &&
-          !normalizedTableName.includes("type") &&
-          !normalizedTableName.includes("log")
-        ) {
-          return await ensureBookings();
-        }
-
-        if (normalizedTableName.includes("usersrights")) {
-          return usersRightsRecords;
-        }
-
-        if (normalizedTableName.includes("usersapprovers")) {
-          return [];
-        }
-
-        if (originalClientFetch) {
-          return await originalClientFetch(tableName, constraints, tenant);
-        }
-
-        return [];
-      };
-
-      const originalGetPaginatedData = (window as any).getPaginatedData;
-
-      (window as any).getPaginatedData = async function (
-        collectionName: string,
-        itemsPerPage: number,
-        filters: any,
-        lastVisible: any,
-        tenant?: string
-      ) {
-        return await ensureBookings();
-      };
-
-      const originalGetDocs = (window as any).getDocs;
-
-      (window as any).getDocs = async function (queryRef: any) {
-        const path = queryRef && (queryRef._path || queryRef.path);
-        if (
-          path &&
-          path.includes("booking") &&
-          !path.includes("bookingTypes") &&
-          !path.includes("bookingLog")
-        ) {
-          const bookings = await ensureBookings();
-          return {
-            docs: bookings.map((booking: any) => ({
-              id: booking.id ?? booking.calendarEventId,
-              data: () => ({ ...booking }),
-            })),
-          };
-        }
-
-        if (path && path.includes("usersRights")) {
-          return {
-            docs: usersRightsRecords.map(
-              (record: any, index: number) => ({
-                id: record.id ?? `users-rights-${index}`,
-                data: () => ({ ...record }),
-              })
-            ),
-          };
-        }
-
-        if (path && path.includes("usersApprovers")) {
-          return {
-            docs: [],
-          };
-        }
-
-        if (originalGetDocs) {
-          return await originalGetDocs(queryRef);
-        }
-
-        return { docs: [] };
-      };
-
-      void loadBookings().catch(() => {});
-
-      const overrideClientGetDataById = async (calendarEventId: string) => {
-        const bookings = await ensureBookings();
-        const match = bookings.find(
-          (booking: any) => booking.calendarEventId === calendarEventId
-        );
-        if (!match) return null;
-        return {
-          id: match.id ?? match.calendarEventId,
-          ...match,
-        };
-      };
-
-      const overrideMap: Record<string, Function | undefined> = {
-        clientFetchAllDataFromCollection: (window as any)
-          .clientFetchAllDataFromCollection,
-        getPaginatedData: (window as any).getPaginatedData,
-        clientGetDataByCalendarEventId: async (
-          _tableName: any,
-          calendarEventId: string,
-          _tenant?: string
-        ) => {
-          return await overrideClientGetDataById(calendarEventId);
-        },
-      };
-
-      const patchWebpackModules = () => {
-        const chunk = (window as any).webpackChunk_N_E;
-        if (!chunk) return false;
-
-        let wpRequire: any;
-        try {
-          chunk.push([
-            ["__e2e_mock_" + Date.now()],
-            {},
-            (req: any) => {
-              wpRequire = req;
-            },
-          ]);
-        } catch (_) {
-          return false;
-        }
-
-        if (!wpRequire?.c) return false;
-
-        let patched = false;
-        Object.values(wpRequire.c).forEach((mod: any) => {
-          if (
-            mod?.exports &&
-            typeof mod.exports === "object" &&
-            mod.exports !== null
-          ) {
-            for (const [key, fn] of Object.entries(overrideMap)) {
-              if (key in mod.exports && fn) {
-                try {
-                  Object.defineProperty(mod.exports, key, {
-                    value: fn,
-                    writable: true,
-                    configurable: true,
-                    enumerable: true,
-                  });
-                  patched = true;
-                } catch (_) {
-                  try {
-                    mod.exports[key] = fn;
-                    patched = true;
-                  } catch (_) {}
-                }
-              }
-            }
-          }
-        });
-        return patched;
-      };
-
-      const _earlyPatchId = setInterval(() => {
-        try {
-          if (patchWebpackModules()) {
-            clearInterval(_earlyPatchId);
-          }
-        } catch (_) {}
-      }, 2);
-      setTimeout(() => clearInterval(_earlyPatchId), 10000);
-
-      try {
-        patchWebpackModules();
-      } catch (_err) {
-        // ignore sync errors
-      }
-
-      (window as any).__applyMockBookingsOverrides = () => {
-        try {
-          patchWebpackModules();
-        } catch (_err) {
-          // ignore sync errors
-        }
-      };
-    },
-    {
-      timestampFields: TIMESTAMP_FIELDS_WITH_BY,
-      initialBooking: basePayload,
-      initialUsersRights: rightsPayload,
-      adminEmail: ADMIN_EMAIL,
-    }
-  );
-}
-
 function updateServicesState(
   existing: any,
   serviceType: ServiceKey,
@@ -774,7 +430,12 @@ test.describe("Full service closeout flow (mocked Firestore)", () => {
 
     await seedAdminUserData();
     await seedCheckedOutBookingWithAllServices();
-    await registerMockBookingsFeed(page);
+    await registerMockBookingsFeed(page, {
+      bookingDocId: BOOKING_DOC_ID,
+      calendarEventId: CALENDAR_EVENT_ID,
+      usersRightsDocId: USERS_RIGHTS_DOC_ID,
+      adminEmail: ADMIN_EMAIL,
+    });
 
     const capturedServiceRequests: any[] = [];
     await mockEndpoints(page, capturedServiceRequests);
